update navigation and file organization
